Added the second copy of updated files
