+++ src/test/java/com/philips/ToolTest.java
@@ -4,10 +4,10 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.assertFalse;
 
 import com.philips.model.Tool;
-import com.philips.action.ExecuteTools;
 import org.junit.Test;
+import java.util.ArrayList;
 
-public class ExecuteToolsTest 
+public class ToolTest 
 {
     //correct
     @Test
@@ -19,7 +19,7 @@ public class ExecuteToolsTest
         toolObj.setoutputFile("outputPMD.txt");
         toolObj.settoolExeFile("C:/Users/320065418/OneDrive - Philips/Desktop/Case Study/FirstProject/staticAnalysisTools/pmd-bin-6.16.0/bin/pmd.bat");
         toolObj.setCommand(exePath);
-        ExecuteTools executeTools=new ExecuteTools();
+        Tool executeTools=new Tool();
         boolean status = executeTools.executetool(toolObj);
         assertTrue("Command run successfully", status );
         System.out.println("Command run successfully");
@@ -34,9 +34,38 @@ public class ExecuteToolsTest
         toolObj.setoutputFile("outputPMD.txt");
         toolObj.settoolExeFile("C:/Users/320065418/OneDrive - Philips/Desktop/Case Study/FirstProject/staticAnalysisTools/pmd-bin-6.16.0/pmd.bat");
         toolObj.setCommand(exePath);
-        ExecuteTools executeTools=new ExecuteTools();
+        Tool executeTools=new Tool();
         boolean status = executeTools.executetool(toolObj);
         assertFalse("Command not run successfully", status );
         System.out.println("Command not run successfully");
     }
+
+    @Test
+    public void congifFileIsNotInSpecifiedFormat() {
+        Tool initializeToolsObj=new Tool();
+
+        ArrayList<Tool> toolobj=initializeToolsObj.initializetool("C:/Users/320065418/OneDrive - Philips/Desktop/Case Study/FirstProject/congif1.properties");
+        boolean status=false;
+
+        if(toolobj!=null)
+            status=true;
+
+        assertFalse("Congif file is not in correct format", status );
+        System.out.println("Config file is not in correct format");
+    }
+
+        //correct
+        @Test
+        public void congifFileIsInSpecifiedFormat() {
+            Tool initializeToolsObj=new Tool();
+    
+            ArrayList<Tool> toolobj=initializeToolsObj.initializetool("C:/Users/320065418/OneDrive - Philips/Desktop/Case Study/FirstProject/config.properties");
+            boolean status=false;
+    
+            if(toolobj!=null)
+                status=true;
+    
+            assertTrue("Config file is in correct format", status );
+            System.out.println("Config file is in correct format");
+        }
 }
